Replace sample data with real data, and add support for categories
Since sample data is no longer being used, there are no restrictions 
placed on the number of resources in any given level of classification, 
so remove the MAX_NUM_RESOURCES_IN_SUBCLASS constant.

The JSON file used to get the Library of Congress Classification 
schedule data does not contain the letters and descriptions for the 
main, top-level classes (A, B, etc.). Create the LCC_MAIN_CLASSES object 
to store this data.

Create an enum-esque chartLevels object to simplify tracking the state 
of the chart.

There are now three possible states of the chart rathe than two--that 
is, there are three possible levels of data: main class, subclass, and 
category. (Note: Here (and throughout the code), I am using the term 
category to describe a third-level classication. For example, for a 
resource with a call number beginning with "AZ231", the main class would 
be "A", the subclass would be "AZ", and I would refer to the category as 
"AZ231".) Because of this, the chartLevel variable and chartLevels 
object are used to track state and the boolean mainClassDataIsDisplayed 
varaible is removed.

Create three Object prototype constructors, one for each of the three 
levels of classification currently supported, To simplify the 
construction of the main data object.

Update the plotly animation functions so reflect the change in data 
source and the added support for categories.

Code generating sample data is no longer neccessary. Remove it.

Update the back button's event listener to account for the added 
possible state of the chart.