Move code that loads and processes the data file
Move the loading and processing of the data file into the callback 
function of the AJAX request for the LCC subclass data, since this data 
is needed to properly and successfully process the data contained in the 
data file. This prevents the data file from potentially being processed 
before the subclass data has loaded.